Renamed to add 'cls' keyword
