+++ conf/script/src/utils/meta_prog/generics/cls_wrapper_data.py
@@ -1,7 +1,7 @@
 import dataclasses
 import typing
 
-from utils.meta_prog.generics.mixin import GenericClassMixin
+from utils.meta_prog.generics.cls_mixin import GenericClassMixin
 
 
 @dataclasses.dataclass()
